requirements.txt
aiohttp>=2.3.10
aiohttp_cors
pycryptodomex
bjson>=0.2.7
nem-ed25519
p2p_python>=1.0.4
dill
psutil
pysha3